Fix bust cache double bust free
Fixed problem where bust and its filename would be freed twice. This is from having a bust fadeout in 0 time because the bust data isn't freed until Update(); is called, which is in the middle of the DrawScene function. So the bust image and path would be added to the cache, freed in the middle of the DrawScene, and then freed next time the cache was freed.
Also remove Windows stuff because too lazy to maintain it.